perf: Optmize querysets to prevent N+1 problem
The API endpoints for 'movies' were suffering from the N+1 query problema, leading to an excessive number of database queries to fetch related data (like actors and genres)

This commit updates the view querysets in `apps/movies/views.py` to use `select_related` (for foreign key relationships) and `prefetch_related` *for many-to-many relationships).

this optimization will significantly redouce the number of queries, impriving the API's response time for movie-related endpoints.